refactor(compositor): move the pure geometry into frame_geometry
`compose_frame` decides two separate things: WHERE each layer goes, and how to hand
that to a GPU. The first is arithmetic — preset placements, source crop, cover-fit,
corner radii, shadow fractions, the fixture timeline, CSS colour parsing — and it
lived in `compositor_windows.rs` purely because that is where the D3D11 engine was
written first.

The macOS port needs the same placements to the pixel — that is the "iso-render"
property the project measures — and the only way two backends genuinely agree is
that they call the same function, not that they maintain two copies obliged to stay
in step. Same argument as `timeline_walk.rs`, one layer up.

So: 337 lines of code and 14 helpers move to `frame_geometry.rs`, declared
`pub mod` with no `cfg`, next to `regions.rs` which is the working precedent for a
portable geometry module. Nothing is rewritten — every moved line is byte-identical
to the original, verified mechanically; the only edits are visibility (`pub(crate)`
so the Windows engine can still reach them, and `pub(crate)` on `Placement` /
`FrameParams` fields, which were module-private and would otherwise become
unreachable) and the imports at both ends.

`compositor_windows.rs` keeps `pub use` on OUT_W / OUT_H / HALF_W / HALF_H /
FIXTURE_FRAMES, so `pipeline_windows.rs`, `live.rs` and `crates/poc-d3d/src/app.rs`
that read them through `crate::compositor::…` need no edit at all.

The immediate payoff: **20 tests that had only ever run on Windows now run on macOS
too** — the crate goes from 73 to 93 on a Mac. They cover exactly the arithmetic the
Metal engine will consume next: `screen_source_rect`, `cover_crop_uv`,
`cover_uv_rect`, `cursor_sprite_dst`, the zoom/padding invariants.

The file-level `allow(dead_code)` is honest rather than defensive: on macOS half of
this module has no caller yet because the Metal engine has no layered `compose_frame`.
It is not dead code, it is code the port has not reached — and it is exercised by its
tests on both platforms. The allow goes away when the layer driving lands.
