+++ crates/compositor/src/frame_geometry.rs
@@ -1,0 +1,882 @@
+//! La géométrie de composition, sans backend.
+//!
+//! Tout ce qui décide OÙ va un calque et de quoi il a l'air — placements de preset,
+//! coupe source, cover-fit, rayons, ombres, timeline de la fixture, parsing des
+//! couleurs CSS — par opposition à ce qui l'envoie au GPU. Rien ici ne connaît
+//! D3D11 ni Metal, et le module est donc compilé sur les deux plateformes
+//! (`pub mod frame_geometry;` sans `cfg`, comme `regions.rs` juste à côté).
+//!
+//! # Pourquoi ce module existe
+//!
+//! Ce code vivait dans `compositor_windows.rs`. Le port macOS a besoin des mêmes
+//! placements au pixel près — c'est la propriété « iso-render » que le projet
+//! mesure — et la seule façon de garantir que deux backends s'accordent est qu'ils
+//! lisent la même fonction, pas qu'ils entretiennent deux copies qui doivent rester
+//! d'accord. C'est le même raisonnement que `timeline_walk.rs`.
+//!
+//! Effet de bord immédiat : cette géométrie et ses tests, qui n'avaient jamais été
+//! exécutés ailleurs que sur Windows, tournent maintenant aussi dans le job macOS.
+
+// Sur macOS, la moitié de ce module est encore sans consommateur : le moteur Metal
+// n'a pas de `compose_frame` en couches, donc rien n'appelle encore `screen_source_rect`,
+// `cover_uv_rect`, les fractions d'ombre ou `CursorPlacement`. Ce n'est PAS du code mort —
+// c'est du code que le port n'a pas encore atteint, et il est exercé par ses tests sur les
+// deux plateformes. Le `allow` saute quand le pilotage des couches arrive côté Metal.
+#![allow(dead_code)]
+
+use crate::config::Cfg;
+use crate::scene::SceneCrop;
+
+pub const OUT_W: u32 = 1920;
+pub const OUT_H: u32 = 1080;
+/// Parse une couleur "#rgb" / "#rrggbb" (sRGB, comme les wallpapers web) → [r,g,b,a] 0..1.
+/// Les couleurs plates suivent le même chemin que `bg_color` (pas de linéarisation).
+/// Décode une data URL base64 (`data:image/png;base64,AAAA…`) en octets. `None` si ce n'en est
+/// pas une — l'appelant retombe alors sur une lecture disque.
+///
+/// Écrit à la main plutôt qu'avec une dépendance : c'est le seul usage de base64 du projet, et le
+/// décodeur tient en quinze lignes vérifiables. Les caractères hors alphabet (retours à la ligne
+/// d'un URI replié, `=` de padding) sont ignorés, ce qui rend la fonction tolérante sans être
+/// laxiste : un caractère invalide ne peut pas décaler le flux, il est simplement absent.
+pub(crate) fn decode_data_uri(uri: &str) -> Option<Vec<u8>> {
+    let rest = uri.strip_prefix("data:")?;
+    let comma = rest.find(',')?;
+    if !rest[..comma].contains("base64") {
+        return None;
+    }
+    let payload = &rest[comma + 1..];
+    let sextet = |c: u8| -> Option<u32> {
+        match c {
+            b'A'..=b'Z' => Some((c - b'A') as u32),
+            b'a'..=b'z' => Some((c - b'a') as u32 + 26),
+            b'0'..=b'9' => Some((c - b'0') as u32 + 52),
+            b'+' => Some(62),
+            b'/' => Some(63),
+            _ => None,
+        }
+    };
+    let mut out = Vec::with_capacity(payload.len() / 4 * 3);
+    let (mut acc, mut bits) = (0u32, 0u32);
+    for byte in payload.bytes() {
+        let Some(v) = sextet(byte) else { continue };
+        acc = (acc << 6) | v;
+        bits += 6;
+        if bits >= 8 {
+            bits -= 8;
+            out.push((acc >> bits) as u8);
+        }
+    }
+    Some(out)
+}
+pub(crate) fn parse_hex(s: &str) -> Option<[f32; 4]> {
+    // Le contrat accepte du CSS, pas seulement de l'hex : la bridge des captions produit du
+    // `rgba(r, g, b, a)` (l'inspector stocke couleur + opacité séparément, et `captionBackgroundCss`
+    // les recombine en rgba pour la preview) et les stops de gradient arrivent aussi sous cette
+    // forme. `transparent` est un cas particulier documenté : alpha 0, pas de plaque. Tout le
+    // reste tombe sur None → l'appelant applique son fallback (alpha 0 pour un fond, alpha 1
+    // pour un texte, etc.) — la même sémantique qu'avant l'ajout du parseur rgba.
+    let trimmed = s.trim();
+    if trimmed.eq_ignore_ascii_case("transparent") {
+        return Some([0.0, 0.0, 0.0, 0.0]);
+    }
+    // CSS Color 4 fait de `rgb()` et `rgba()` des synonymes : les deux acceptent 3 ou 4
+    // composantes. On les traite donc par le même chemin plutôt que d'imposer une arité par
+    // nom — refuser `rgba(0, 0, 0)` ne « signalerait » rien d'utile, ça retomberait sur le
+    // fallback de l'appelant, c'est-à-dire une plaque invisible : exactement le bug #178.
+    if let Some(inner) =
+        strip_color_fn(trimmed, "rgba").or_else(|| strip_color_fn(trimmed, "rgb"))
+    {
+        return parse_rgb_components(inner);
+    }
+    let h = trimmed.trim_start_matches('#');
+    // Un corps hex est ASCII par définition, et les découpes par octet ci-dessous (`h[i..=i]`,
+    // `h[0..2]`…) paniqueraient au milieu d'un caractère multi-octets qui ferait pile 3 ou 6
+    // octets (`éa`, `€€`). On refuse avant de découper.
+    if !h.is_ascii() {
+        return None;
+    }
+    let (r, g, b) = match h.len() {
+        3 => {
+            let d = |i: usize| u8::from_str_radix(&h[i..=i], 16).ok().map(|v| v * 17);
+            (d(0)?, d(1)?, d(2)?)
+        }
+        6 => (
+            u8::from_str_radix(&h[0..2], 16).ok()?,
+            u8::from_str_radix(&h[2..4], 16).ok()?,
+            u8::from_str_radix(&h[4..6], 16).ok()?,
+        ),
+        _ => return None,
+    };
+    Some([r as f32 / 255.0, g as f32 / 255.0, b as f32 / 255.0, 1.0])
+}
+/// `rgba(0, 0, 0, 0.55)` → `"0, 0, 0, 0.55"` (le contenu entre les parenthèses), None si
+/// l'enveloppe n'est pas de la forme `fn(...)`. Tolère les espaces et les tabs, refuse les
+/// virgules finales et les arguments vides — le gradient parser a déjà démontré que la couche
+/// application produit des chaînes propres, donc rester strict ici évite d'avaler des CSS
+/// tordus qu'on ne maîtrise pas. La casse du préfixe est libre (`RGBA(...)` est valide) parce
+/// que CSS le permet.
+pub(crate) fn strip_color_fn<'a>(s: &'a str, name: &str) -> Option<&'a str> {
+    // `get` rend None si `name.len()` n'est pas une frontière de caractère : c'est ce qui rend
+    // le slice `s[..name.len()]` juste en dessous sûr par construction. Un `&s[..n]` direct
+    // paniquerait au milieu d'un caractère multi-octets (`#ab€cd` coupe dans le `€`), et une
+    // panique traverserait le pont N-API au lieu de retomber sur le fallback de l'appelant —
+    // le contraire de ce que ce parseur promet.
+    let after_name = s.get(name.len()..)?;
+    if !s[..name.len()].eq_ignore_ascii_case(name) {
+        return None;
+    }
+    let inner = after_name.strip_prefix('(')?.strip_suffix(')')?.trim();
+    if inner.is_empty() {
+        return None;
+    }
+    Some(inner)
+}
+/// `"r, g, b"` ou `"r, g, b, a"` (floats 0..255 pour r/g/b, 0..1 pour a) → `[r, g, b, a]` en
+/// 0..1, l'alpha valant 1 (opaque) quand elle est absente. Toute autre arité → None. Tolère
+/// les espaces autour des virgules, pas les pourcentages : le gradient parser n'envoie pas de
+/// `rgb(50%, …)` et les couches UI qui le font n'arrivent pas ici (les couleurs wallpaper
+/// passent par une autre route, cf. `parseWallpaper`).
+pub(crate) fn parse_rgb_components(s: &str) -> Option<[f32; 4]> {
+    let parts: Vec<&str> = s.split(',').map(str::trim).collect();
+    let (rgb, alpha) = match parts.as_slice() {
+        [r, g, b] => ([r, g, b], 1.0),
+        // L'alpha est déjà sur [0..1] par convention (`rgba(...,0.55)`, pas `rgba(...,55)`).
+        [r, g, b, a] => ([r, g, b], parse_color_channel(a, 1.0)?),
+        _ => return None,
+    };
+    Some([
+        parse_color_channel(rgb[0], 255.0)?,
+        parse_color_channel(rgb[1], 255.0)?,
+        parse_color_channel(rgb[2], 255.0)?,
+        alpha,
+    ])
+}
+pub(crate) fn parse_color_channel(raw: &str, max: f32) -> Option<f32> {
+    let n: f32 = raw.parse().ok()?;
+    if !n.is_finite() || n < 0.0 || n > max {
+        return None;
+    }
+    Some(n / max)
+}
+/// Rect source après crop puis zoom, dans les UV de la texture D3D. `u_max`/`v_max`
+/// excluent le padding NV12 ; le crop reste donc exprimé dans le frame visible (0..1),
+/// comme `VirtualPreview.cropVideoStyle`, puis le focus du zoom est remappé dans ce crop.
+pub(crate) fn screen_source_rect(
+    u_max: f32,
+    v_max: f32,
+    crop: Option<SceneCrop>,
+    zoom: f32,
+    focus: [f32; 2],
+) -> [f32; 4] {
+    let normalized_crop = crop.and_then(|crop| {
+        if !crop.x.is_finite() || !crop.y.is_finite()
+            || !crop.width.is_finite() || !crop.height.is_finite()
+        {
+            return None;
+        }
+        let x0 = crop.x.clamp(0.0, 1.0);
+        let y0 = crop.y.clamp(0.0, 1.0);
+        let x1 = (crop.x + crop.width).clamp(x0, 1.0);
+        let y1 = (crop.y + crop.height).clamp(y0, 1.0);
+        (x1 > x0 && y1 > y0).then_some([x0, y0, x1, y1])
+    });
+    let [x0, y0, x1, y1] = normalized_crop.unwrap_or([0.0, 0.0, 1.0, 1.0]);
+    let (cu0, cv0, cu1, cv1) = (x0 * u_max, y0 * v_max, x1 * u_max, y1 * v_max);
+    let (cw, ch) = (cu1 - cu0, cv1 - cv0);
+    let zoom = if zoom.is_finite() && zoom >= 1.0 { zoom } else { 1.0 };
+    let fx = if focus[0].is_finite() { focus[0].clamp(0.0, 1.0) } else { 0.5 };
+    let fy = if focus[1].is_finite() { focus[1].clamp(0.0, 1.0) } else { 0.5 };
+    let (hu, hv) = (cw / (2.0 * zoom), ch / (2.0 * zoom));
+    // `.max(cu0/cv0)` absorbs the tiny float inversion possible at zoom=1.
+    let su0 = (cu0 + fx * cw - hu).clamp(cu0, (cu1 - 2.0 * hu).max(cu0));
+    let sv0 = (cv0 + fy * ch - hv).clamp(cv0, (cv1 - 2.0 * hv).max(cv0));
+    [su0, sv0, su0 + 2.0 * hu, sv0 + 2.0 * hv]
+}
+/// Rect DESTINATION de l'écran quand on dessine une coupe source PLUS LARGE que celle qui
+/// remplissait la boîte — le cœur du correctif #179.
+///
+/// Le zoom natif se jouait entièrement dans la coupe source (`screen_source_rect` rétrécit
+/// la coupe autour du focus) pendant que la boîte, elle, ne bougeait pas : le zoom
+/// s'arrêtait donc à la frontière paddée au lieu d'atteindre les bords du cadre. La
+/// référence fait l'inverse — `applyZoomTransform` (TS) met à l'échelle et translate le
+/// CONTENEUR CAMÉRA, masque compris, donc la boîte paddée grandit avec le zoom, sort de
+/// l'étage, et le padding s'efface.
+///
+/// On rend donc le zoom à la boîte : la coupe dessinée redevient le simple crop
+/// (`cut`, zoom 1) et c'est la boîte qui porte le grossissement. `cut_ref` est la coupe
+/// d'AVANT (zoom entier, celle qui remplissait `base`) et sert de référence : on reporte
+/// `cut` à travers le mapping `cut_ref → base`.
+///
+/// C'est ce report qui fait toute la sûreté du correctif. Le mapping image→écran est
+/// conservé PAR CONSTRUCTION — même grossissement, même cadrage, même point de focus au
+/// même pixel — quel que soit le crop, le clamp de bord ou le `cover`, puisque tout cela
+/// est déjà cuit dans les deux coupes. Seule l'ÉTENDUE dessinée grandit, et c'est
+/// exactement elle qui déborde le padding. Tout ce qui roule sur ce mapping (curseur,
+/// tilt 3D, motion blur) est donc inchangé.
+///
+/// Pas de clamp dans le cadre : la boîte doit pouvoir en sortir (« No stage clamping »,
+/// `frameRenderer.cameraAwareMaskRect`) — le rasterizer coupe ce qui dépasse, comme il le
+/// fait déjà pour le fond flouté.
+pub(crate) fn remap_box(base: [f32; 4], cut_ref: [f32; 4], cut: [f32; 4]) -> [f32; 4] {
+    let (rw, rh) = ((cut_ref[2] - cut_ref[0]), (cut_ref[3] - cut_ref[1]));
+    if !(rw > 1e-6 && rh > 1e-6) {
+        return base;
+    }
+    [
+        base[0] + base[2] * (cut[0] - cut_ref[0]) / rw,
+        base[1] + base[3] * (cut[1] - cut_ref[1]) / rh,
+        base[2] * (cut[2] - cut[0]) / rw,
+        base[3] * (cut[3] - cut[1]) / rh,
+    ]
+}
+/// Sous-rect SOURCE (en UV de texture) qui remplit une boîte de ratio `box_ar` **sans
+/// déformer** l'image : le plus grand rect centré ayant ce ratio, tiré de la frame
+/// visible — l'équivalent de `object-fit: cover` côté web.
+///
+/// C'est LA primitive qui garantit qu'une couche vidéo n'est jamais étirée. Le
+/// contrat est déplacé de l'appelant (« donne-moi un dst au ratio de la source »,
+/// hypothèse qu'un preset pouvait violer en silence) vers le calcul lui-même
+/// (« quel que soit le dst, je choisis la coupe qui l'habille »).
+///
+/// * `visible` : dimensions RÉELLES de l'image dans la texture (`AVFrame::width/height`) ;
+///   elles peuvent être plus petites que la texture, qui est allouée avec du padding
+///   décodeur — d'où la division finale par `tex`.
+/// * `tex` : dimensions de la texture, pour normaliser en UV.
+/// * `box_ar` : ratio largeur/hauteur de la boîte de destination, en pixels de rendu.
+///
+/// Retourne `(u0, v0, u1, v1)`. Quand la boîte a déjà le ratio de la source, la coupe
+/// est la frame entière — donc aucun changement de pixel sur les placements qui étaient
+/// déjà corrects.
+pub(crate) fn cover_crop_uv(visible: [f32; 2], tex: [f32; 2], box_ar: f32) -> (f32, f32, f32, f32) {
+    let (cam_w, cam_h) = (visible[0].max(1.0), visible[1].max(1.0));
+    let (tex_w, tex_h) = (tex[0].max(1.0), tex[1].max(1.0));
+    let full = [0.0, 0.0, cam_w / tex_w, cam_h / tex_h];
+    let [u0, v0, u1, v1] = cover_uv_rect(full, tex, box_ar);
+    (u0, v0, u1, v1)
+}
+/// Rétrécit un rect SOURCE déjà exprimé en UV (`[u0, v0, u1, v1]`) autour de son
+/// centre pour qu'il porte le ratio `box_ar` une fois rapporté aux pixels de la
+/// texture. C'est la forme générale de `object-fit: cover`, et LA primitive qui
+/// garantit qu'une couche vidéo n'est jamais étirée.
+///
+/// Deux appelants, deux points d'entrée dans le rect :
+///   - la **webcam** part de la frame visible entière (`cover_crop_uv`) ;
+///   - l'**écran** part du rect déjà réduit par le crop utilisateur ET le zoom,
+///     et n'applique ce cover que dans les layouts qui le demandent
+///     (`Scene.layout.screen_cover` — les blocs side-by-side / top-bottom, où le
+///     web fait exactement la même chose via `screenCover`).
+///
+/// Rogner APRÈS le crop et le zoom est ce qui rend l'opération composable : le
+/// crop décide quoi montrer, le zoom où regarder, le cover comment habiller la
+/// boîte. Chacun réduit le rect précédent, jamais ne le déforme.
+///
+/// Quand le rect a déjà le ratio de la boîte, il est renvoyé inchangé — donc
+/// aucun placement déjà correct ne bouge.
+pub(crate) fn cover_uv_rect(uv: [f32; 4], tex: [f32; 2], box_ar: f32) -> [f32; 4] {
+    let (tex_w, tex_h) = (tex[0].max(1.0), tex[1].max(1.0));
+    let (w_uv, h_uv) = ((uv[2] - uv[0]).max(1e-6), (uv[3] - uv[1]).max(1e-6));
+    // ratio du rect courant, en PIXELS (les UV sont anisotropes dès que la
+    // texture n'est pas carrée — d'où le passage par `tex`).
+    let (w_px, h_px) = (w_uv * tex_w, h_uv * tex_h);
+    let cur_ar = w_px / h_px;
+    let box_ar = if box_ar.is_finite() && box_ar > 0.0 { box_ar } else { cur_ar };
+    let (new_w_px, new_h_px) = if box_ar >= cur_ar {
+        (w_px, w_px / box_ar) // boîte plus large → pleine largeur, on rogne en hauteur
+    } else {
+        (h_px * box_ar, h_px) // boîte plus haute → pleine hauteur, on rogne en largeur
+    };
+    let (new_w, new_h) = (new_w_px / tex_w, new_h_px / tex_h);
+    let (cx, cy) = (uv[0] + w_uv * 0.5, uv[1] + h_uv * 0.5);
+    [cx - new_w * 0.5, cy - new_h * 0.5, cx + new_w * 0.5, cy + new_h * 0.5]
+}
+pub const HALF_W: u32 = OUT_W / 2;
+pub const HALF_H: u32 = OUT_H / 2;
+pub const FIXTURE_FRAMES: u32 = 360;
+pub(crate) const FPS: f32 = 60.0;
+/// Longueurs de style exprimées en FRACTION du petit côté du cadre, et non en pixels.
+///
+/// Elles étaient écrites en px bruts au point d'appel, ce qui voulait dire « px du render
+/// target » — donc une proportion DIFFÉRENTE selon la taille de rendu : 40 px, c'est 3,7 % d'un
+/// cadre 1080 mais 1,9 % d'un 2160. L'ombre était donc deux fois plus douce en preview qu'à
+/// l'export, et un export 4K la recevait deux fois plus faible qu'un 1080p — même famille de bug
+/// que les rayons venus de l'app, mais née à l'intérieur du natif. Les valeurs ci-dessous sont
+/// les anciennes constantes rapportées au cadre 1080 contre lequel elles avaient été réglées :
+/// le rendu à cette résolution est donc inchangé, et devient enfin identique partout ailleurs.
+pub(crate) const SHADOW_TUNING_REF_PX: f32 = 1080.0;
+pub(crate) const SCREEN_SHADOW_SPREAD_FRAC: f32 = 40.0 / SHADOW_TUNING_REF_PX;
+pub(crate) const SCREEN_SHADOW_OFFSET_FRAC: f32 = 16.0 / SHADOW_TUNING_REF_PX;
+pub(crate) const WEBCAM_SHADOW_SPREAD_FRAC: f32 = 32.0 / SHADOW_TUNING_REF_PX;
+pub(crate) const WEBCAM_SHADOW_OFFSET_FRAC: f32 = 12.0 / SHADOW_TUNING_REF_PX;
+/// Opacité FIXE de l'ombre portée de la caméra (layout PiP uniquement). Contrairement à
+/// l'ombre de l'écran — dont l'opacité est pilotée par le slider Shadow (`shadow_scale`) —
+/// l'ombre de la caméra est une ombre légère NON paramétrable : même valeur quelle que soit
+/// la position du slider. Parité avec le preset PiP côté web (`compositeLayout.ts`,
+/// `rgba(0,0,0,0.35)`), dont l'ombre est elle aussi un forfait fixe et PiP-only.
+pub(crate) const WEBCAM_SHADOW_OPACITY: f32 = 0.35;
+/// Taille de base du curseur, même convention (34 px réglés contre un cadre 1080).
+pub(crate) const CURSOR_BASE_SIZE_FRAC: f32 = 34.0 / SHADOW_TUNING_REF_PX;
+/// Rect [x,y,w,h] normalisé d'un sprite de curseur de taille `w`×`h` dont le pivot `hotspot`
+/// (fraction 0..1 de l'image) doit tomber exactement sur `center`.
+///
+/// L'invariant est que `center` reste sur le pixel désigné QUELLE QUE SOIT la taille : le
+/// décalage grandit avec le sprite, donc il doit être une fraction de `w`/`h` et pas une
+/// constante. Un pivot centré en dur (0.5) laissait la pointe dériver de plus en plus loin de
+/// la zone visée à mesure qu'on agrandissait le curseur.
+pub(crate) fn cursor_sprite_dst(center: [f32; 2], w: f32, h: f32, hotspot: [f32; 2]) -> [f32; 4] {
+    [center[0] - w * hotspot[0], center[1] - h * hotspot[1], w, h]
+}
+/// Où poser le curseur, et dans quel repère.
+///
+/// Le curseur remplace un pointeur qui faisait partie de l'image capturée, donc il vit SUR la
+/// surface de l'écran, pas dans un calque au-dessus. Quand cet écran est incliné en 3D, ce n'est
+/// donc pas seulement sa position qu'il faut projeter mais son sprite entier : autrement il se
+/// lit comme un autocollant plat posé sur une scène en perspective.
+#[derive(Clone, Copy)]
+pub(crate) enum CursorPlacement {
+    /// Écran droit : centre en coordonnées sortie 0..1.
+    Upright { center: [f32; 2] },
+    /// Écran incliné : position 0..1 DANS le plan, plus de quoi projeter les coins du sprite.
+    Tilted {
+        /// Position du pivot dans le plan (0..1 depuis son coin haut-gauche).
+        plane_pt: [f32; 2],
+        quad: crate::regions::TiltedQuad,
+        /// Centre du plan en px sortie — `quad.corners` y est relatif.
+        center_px: [f32; 2],
+        /// Taille du rect d'écran NON incliné en px : l'unité dans laquelle la taille du
+        /// curseur est exprimée, et donc ce qui la convertit en fraction du plan.
+        screen_px: [f32; 2],
+        /// Taille de la cible de rendu en px, pour repasser des px aux 0..1 de la sortie.
+        render_px: [f32; 2],
+    },
+}
+impl CursorPlacement {
+    /// Interpolation entre deux placements, pour les copies de la traînée de flou. Sur un plan
+    /// incliné on interpole DANS le plan : la traînée suit alors la surface au lieu de couper
+    /// droit à travers la perspective.
+    pub(crate) fn lerp(self, other: CursorPlacement, f: f32) -> CursorPlacement {
+        match (self, other) {
+            (
+                CursorPlacement::Tilted { plane_pt: a, quad, center_px, screen_px, render_px },
+                CursorPlacement::Tilted { plane_pt: b, .. },
+            ) => CursorPlacement::Tilted {
+                plane_pt: [a[0] + (b[0] - a[0]) * f, a[1] + (b[1] - a[1]) * f],
+                quad,
+                center_px,
+                screen_px,
+                render_px,
+            },
+            (a, b) => {
+                let (p, q) = (a.upright_center(), b.upright_center());
+                CursorPlacement::Upright {
+                    center: [p[0] + (q[0] - p[0]) * f, p[1] + (q[1] - p[1]) * f],
+                }
+            }
+        }
+    }
+
+    /// Le centre en coordonnées sortie, quel que soit le repère — ce dont ont besoin le curseur
+    /// math de secours et le calcul de vélocité.
+    pub(crate) fn upright_center(self) -> [f32; 2] {
+        match self {
+            CursorPlacement::Upright { center } => center,
+            CursorPlacement::Tilted { plane_pt, quad, center_px, render_px, .. } => {
+                let (px, py) = quad.point_px(plane_pt[0], plane_pt[1]);
+                [(center_px[0] + px) / render_px[0], (center_px[1] + py) / render_px[1]]
+            }
+        }
+    }
+}
+pub(crate) fn ease_in_out_cubic(x: f32) -> f32 {
+    let x = x.clamp(0.0, 1.0);
+    if x < 0.5 {
+        4.0 * x * x * x
+    } else {
+        1.0 - (-2.0 * x + 2.0).powi(3) / 2.0
+    }
+}
+pub(crate) fn lerp(a: f32, b: f32, t: f32) -> f32 {
+    a + (b - a) * t
+}
+pub(crate) fn lerp4(a: [f32; 4], b: [f32; 4], t: f32) -> [f32; 4] {
+    [lerp(a[0], b[0], t), lerp(a[1], b[1], t), lerp(a[2], b[2], t), lerp(a[3], b[3], t)]
+}
+/// Un calque vidéo animé (rect sortie, taille px, rayon) — screen ou webcam.
+#[derive(Clone, Copy)]
+pub(crate) struct Placement {
+    pub(crate) dst: [f32; 4],
+    pub(crate) radius: f32,
+}
+/// Paramètres d'une frame : dérivés du temps par la timeline (§8).
+#[derive(Clone, Copy)]
+pub(crate) struct FrameParams {
+    pub(crate) zoom: f32,
+    pub(crate) focus: [f32; 2],
+    pub(crate) screen: Placement,
+    pub(crate) webcam: Placement, // dst carré (w en px via OUT_W)
+}
+/// Timeline figée de la fixture (6 s) : zoom 1.0→1.8→1.0, layout A(PIP)↔B(côte à côte).
+/// `frame` fractionnaire pour permettre le supersampling temporel (flou de mouvement).
+/// Gaté par `cfg` : zoom et layout ne bougent que si activés.
+pub(crate) fn timeline(frame: f32, cfg: &Cfg) -> FrameParams {
+    let t = frame / FPS; // secondes
+
+    // zoom : montée [0,3s] puis descente [3s,6s], easeInOutCubic
+    let zoom = if cfg.zoom {
+        let zt = if t < 3.0 { ease_in_out_cubic(t / 3.0) } else { ease_in_out_cubic((6.0 - t) / 3.0) };
+        1.0 + 0.8 * zt
+    } else {
+        1.0
+    };
+
+    // layout A = PIP bas-droite ; B = côte à côte. Transitions A→B [2,2.5]s, B→A [4,4.5]s.
+    let lf = if !cfg.layout_anim {
+        0.0
+    } else if t < 2.0 {
+        0.0
+    } else if t < 2.5 {
+        ease_in_out_cubic((t - 2.0) / 0.5)
+    } else if t < 4.0 {
+        1.0
+    } else if t < 4.5 {
+        1.0 - ease_in_out_cubic((t - 4.0) / 0.5)
+    } else {
+        0.0
+    };
+
+    // Layout A (PIP)
+    let a_screen = Placement { dst: [0.05, 0.05, 0.90, 0.90], radius: 24.0 };
+    let a_side = 320.0_f32;
+    let a_webcam = Placement {
+        dst: [
+            (OUT_W as f32 - 40.0 - a_side) / OUT_W as f32,
+            (OUT_H as f32 - 40.0 - a_side) / OUT_H as f32,
+            a_side / OUT_W as f32,
+            a_side / OUT_H as f32,
+        ],
+        radius: 40.0,
+    };
+    // Layout B (côte à côte) : screen à gauche (16:9), webcam carré à droite
+    let b_screen = Placement { dst: [0.035, 0.22, 0.60, 0.5625], radius: 20.0 };
+    let b_side = 520.0_f32;
+    let b_webcam = Placement {
+        dst: [
+            0.70,
+            (OUT_H as f32 - b_side) * 0.5 / OUT_H as f32,
+            b_side / OUT_W as f32,
+            b_side / OUT_H as f32,
+        ],
+        radius: 40.0,
+    };
+
+    FrameParams {
+        zoom,
+        focus: [0.5, 0.32],
+        screen: Placement { dst: lerp4(a_screen.dst, b_screen.dst, lf), radius: lerp(a_screen.radius, b_screen.radius, lf) },
+        webcam: Placement { dst: lerp4(a_webcam.dst, b_webcam.dst, lf), radius: lerp(a_webcam.radius, b_webcam.radius, lf) },
+    }
+}
+/// Placements statiques screen+webcam pour un preset de layout de l'app (contrat de scène) —
+/// remplace le planning A↔B fixture de `timeline()`. Zoom = 1 (les zoom regions viennent ensuite).
+/// La taille/forme/miroir webcam restent appliqués par-dessus via `LiveParams`.
+pub(crate) fn preset_placements(preset: &str) -> FrameParams {
+    // plein cadre : le padding l'insère ensuite (padding 0 → bord à bord).
+    let full_screen = Placement { dst: [0.0, 0.0, 1.0, 1.0], radius: 24.0 };
+    // PiP bas-droite (≈ layout A fixture).
+    let a_side = 320.0_f32;
+    let pip_webcam = Placement {
+        dst: [
+            (OUT_W as f32 - 40.0 - a_side) / OUT_W as f32,
+            (OUT_H as f32 - 40.0 - a_side) / OUT_H as f32,
+            a_side / OUT_W as f32,
+            a_side / OUT_H as f32,
+        ],
+        radius: 40.0,
+    };
+    // webcam hors écran (no-webcam) : quad de taille nulle, jamais visible.
+    let off_webcam = Placement { dst: [2.0, 2.0, 0.0, 0.0], radius: 0.0 };
+
+    let (screen, webcam) = match preset {
+        "dual-frame" => {
+            // côte à côte : screen 16:9 à gauche, webcam carré à droite (≈ layout B fixture).
+            let b_side = 520.0_f32;
+            (
+                Placement { dst: [0.035, 0.22, 0.60, 0.5625], radius: 20.0 },
+                Placement {
+                    dst: [
+                        0.70,
+                        (OUT_H as f32 - b_side) * 0.5 / OUT_H as f32,
+                        b_side / OUT_W as f32,
+                        b_side / OUT_H as f32,
+                    ],
+                    radius: 40.0,
+                },
+            )
+        }
+        "vertical-stack" => {
+            // haut/bas : screen en haut, webcam carré centré en bas.
+            let w_side = 360.0_f32;
+            (
+                Placement { dst: [0.13, 0.04, 0.74, 0.52], radius: 20.0 },
+                Placement {
+                    dst: [
+                        0.5 - (w_side * 0.5) / OUT_W as f32,
+                        0.60,
+                        w_side / OUT_W as f32,
+                        w_side / OUT_H as f32,
+                    ],
+                    radius: 40.0,
+                },
+            )
+        }
+        "no-webcam" => (full_screen, off_webcam),
+        _ => (full_screen, pip_webcam), // "picture-in-picture" (défaut)
+    };
+
+    FrameParams { zoom: 1.0, focus: [0.5, 0.5], screen, webcam }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    /// Le pivot doit rester collé à `center` quand le sprite grandit — c'est exactement ce qui
+    /// était cassé (ancrage centré en dur : la pointe s'éloignait proportionnellement à la
+    /// taille). On dessine la même flèche à deux tailles et on vérifie que le point désigné
+    /// ne bouge pas.
+    #[test]
+    fn sprite_hotspot_stays_on_target_at_any_size() {
+        let center = [0.4, 0.6];
+        let hotspot = [0.119, 0.0874]; // flèche intégrée : la pointe, près du coin haut-gauche
+
+        for (w, h) in [(0.02, 0.04), (0.08, 0.16)] {
+            let dst = cursor_sprite_dst(center, w, h, hotspot);
+            let pivot = [dst[0] + dst[2] * hotspot[0], dst[1] + dst[3] * hotspot[1]];
+            assert!((pivot[0] - center[0]).abs() < 1e-6, "x drifted at {w}x{h}: {pivot:?}");
+            assert!((pivot[1] - center[1]).abs() < 1e-6, "y drifted at {w}x{h}: {pivot:?}");
+            assert_eq!([dst[2], dst[3]], [w, h], "taille altérée");
+        }
+
+        // Et un pivot centré reste bien l'ancien comportement, pour les sprites qui le veulent
+        // (viseur, I-beam, poignées de redimensionnement).
+        assert_eq!(cursor_sprite_dst([0.5, 0.5], 0.2, 0.2, [0.5, 0.5]), [0.4, 0.4, 0.2, 0.2]);
+    }
+    fn assert_rect(actual: [f32; 4], expected: [f32; 4]) {
+        for (actual, expected) in actual.into_iter().zip(expected) {
+            assert!((actual - expected).abs() < 1e-6, "actual={actual}, expected={expected}");
+        }
+    }
+
+    #[test]
+    fn decodes_a_base64_data_uri() {
+        // "Hi!" -> SGkh
+        assert_eq!(decode_data_uri("data:image/png;base64,SGkh").unwrap(), b"Hi!".to_vec());
+    }
+
+    /// L'inspector stocke les couleurs de caption comme `couleur_hex` + `opacité` puis la
+    /// bridge JS recombine en `rgba(r, g, b, a)` pour la preview. Le natif doit rendre la même
+    /// plaque (couleur et opacité) — sinon le calque disparaît silencieusement et la caption
+    /// n'apparaît qu'en texte brut dans l'export. C'était exactement le bug de l'issue #178.
+    #[test]
+    fn parse_hex_understands_rgba_caption_backgrounds() {
+        let parsed = parse_hex("rgba(0, 0, 0, 0.55)").expect("rgba doit parser");
+        assert!((parsed[3] - 0.55).abs() < 1e-6, "alpha 0.55 transmise, pas tombée à 0");
+        assert_eq!([parsed[0], parsed[1], parsed[2]], [0.0, 0.0, 0.0]);
+    }
+
+    /// `rgb(...)` sans alpha est sémantiquement `rgba(..., 1)` — il faut le supporter pour
+    /// qu'un inspector qui n'expose pas d'opacité n'écrive pas un fond invisible.
+    #[test]
+    fn parse_hex_treats_rgb_as_opaque() {
+        let parsed = parse_hex("rgb(255, 128, 0)").expect("rgb doit parser");
+        assert_eq!(parsed, [1.0, 128.0 / 255.0, 0.0, 1.0]);
+    }
+
+    /// Le cas "transparent" est documenté dans le code d'appel : on garde la sémantique
+    /// historique (alpha 0) — la plaque est sautée côté rastérisation, ce qui est exactement ce
+    /// que veut le CSS. Le nouveau parseur ne doit pas le casser.
+    #[test]
+    fn parse_hex_keeps_transparent_at_alpha_zero() {
+        assert_eq!(parse_hex("transparent"), Some([0.0, 0.0, 0.0, 0.0]));
+        // La casse ne doit pas non plus casser : CSS autorise `TRANSPARENT` en théorie, et
+        // refuse une chaîne qui ressemble à un rgba mal formé.
+        assert_eq!(parse_hex("Transparent"), Some([0.0, 0.0, 0.0, 0.0]));
+        assert_eq!(parse_hex("rgba(0, 0, 0, 0)"), Some([0.0, 0.0, 0.0, 0.0]));
+    }
+
+    /// Le contrat historique `#rrggbb` / `rrggbb` ne doit pas régresser : les annotations
+    /// normales (saisies via `ColorField`) ne passent que par ce chemin, et leurs snapshots
+    /// ne pardonneraient pas un changement d'alpha implicite.
+    #[test]
+    fn parse_hex_still_understands_hex_colours() {
+        assert_eq!(parse_hex("#fff"), Some([1.0, 1.0, 1.0, 1.0]));
+        assert_eq!(parse_hex("#000000"), Some([0.0, 0.0, 0.0, 1.0]));
+        assert_eq!(
+            parse_hex("ff8800"),
+            Some([1.0, 136.0 / 255.0, 0.0, 1.0])
+        );
+    }
+
+    /// Hors-format (channel > 255, chaîne vide, named color) → None → l'appelant retombe sur
+    /// son fallback. C'est la même politique qu'avant l'ajout du parseur rgba, on la garde
+    /// explicite pour qu'elle ne dérive pas.
+    #[test]
+    fn parse_hex_rejects_malformed_colours() {
+        assert_eq!(parse_hex(""), None);
+        assert_eq!(parse_hex("not-a-color"), None);
+        assert_eq!(parse_hex("rgba(256, 0, 0, 1)"), None); // canal >255
+        assert_eq!(parse_hex("rgba(0, 0, 0, 1.5)"), None); // alpha >1
+        assert_eq!(parse_hex("rgba(0, 0, 0, 0.5, 1)"), None); // 5 composantes
+        assert_eq!(parse_hex("rgb(0, 0)"), None); // 2 composantes
+    }
+
+    /// CSS Color 4 : `rgb()` et `rgba()` sont synonymes, les deux prennent 3 ou 4 composantes.
+    /// Une couleur bien formée ne doit pas finir sur le fallback de l'appelant — pour un fond
+    /// c'est alpha 0, donc une plaque invisible, soit très exactement le symptôme de #178.
+    #[test]
+    fn parse_hex_accepts_both_arities_on_both_names() {
+        assert_eq!(parse_hex("rgba(0, 0, 0)"), Some([0.0, 0.0, 0.0, 1.0]));
+        assert_eq!(parse_hex("rgb(0, 0, 0, 0.5)"), Some([0.0, 0.0, 0.0, 0.5]));
+    }
+
+    /// Une couleur non-ASCII doit être refusée, pas paniquer : `strip_color_fn` découpait
+    /// `s[..3]` / `s[..4]` sans vérifier la frontière de caractère, donc `#ab€cd` (le `€` occupe
+    /// les octets 3..6) tuait le process au lieu de retomber sur le fallback. `parseWallpaper`
+    /// laisse passer n'importe quelle chaîne préfixée `#` jusqu'ici, une panique côté natif
+    /// traverserait le pont N-API et emporterait l'export.
+    #[test]
+    fn parse_hex_refuses_non_ascii_without_panicking() {
+        assert_eq!(parse_hex("#ab€cd"), None);
+        assert_eq!(parse_hex("rg€(0, 0, 0)"), None);
+        assert_eq!(parse_hex("é"), None);
+        assert_eq!(parse_hex("🎨🎨"), None);
+        // Le chemin hex découpe par octet sur les longueurs 3 et 6 : `éa` fait 3 octets et
+        // `€€` en fait 6, donc les deux tombaient pile sur une découpe intra-caractère.
+        assert_eq!(parse_hex("éa"), None);
+        assert_eq!(parse_hex("€€"), None);
+    }
+
+    #[test]
+    fn ignores_padding_and_line_breaks_inside_the_payload() {
+        // Un URI replié ou paddé doit décoder à l'identique : les caractères hors alphabet sont
+        // sautés, donc ils ne peuvent pas décaler le flux.
+        let folded = "data:image/png;base64,SGkh
+==";
+        assert_eq!(decode_data_uri(folded).unwrap(), b"Hi!".to_vec());
+    }
+
+    #[test]
+    fn a_plain_path_is_not_a_data_uri() {
+        // Le repli lecture-disque des wallpapers en dépend.
+        assert!(decode_data_uri("/wallpapers/x.jpg").is_none());
+        assert!(decode_data_uri("C:/img/y.png").is_none());
+    }
+
+    #[test]
+    fn a_non_base64_data_uri_is_refused() {
+        // `data:image/svg+xml,<svg…>` n'est pas du base64 : mieux vaut échouer que décoder du
+        // texte comme des octets.
+        assert!(decode_data_uri("data:image/svg+xml,<svg/>").is_none());
+    }
+
+    #[test]
+    fn crop_maps_visible_frame_fractions_to_texture_uvs() {
+        let crop = SceneCrop { x: 0.25, y: 0.1, width: 0.5, height: 0.6 };
+        assert_rect(screen_source_rect(0.8, 0.9, None, 1.0, [0.2, 0.7]), [0.0, 0.0, 0.8, 0.9]);
+        assert_rect(screen_source_rect(0.8, 0.9, Some(crop), 1.0, [0.5, 0.5]), [0.2, 0.09, 0.6, 0.63]);
+    }
+
+    #[test]
+    fn zoom_focus_is_applied_inside_the_crop() {
+        let crop = SceneCrop { x: 0.25, y: 0.1, width: 0.5, height: 0.6 };
+        assert_rect(screen_source_rect(0.8, 0.9, Some(crop), 2.0, [0.5, 0.5]), [0.3, 0.225, 0.5, 0.495]);
+        assert_rect(screen_source_rect(0.8, 0.9, Some(crop), 2.0, [1.0, 1.0]), [0.4, 0.36, 0.6, 0.63]);
+    }
+
+    // --- le zoom rendu à la boîte (issue #179) ------------------------------
+    // Le zoom déplace et agrandit la boîte au lieu de rétrécir la coupe. Deux choses à
+    // figer, et elles tirent en sens inverse : la boîte DOIT déborder le padding (l'issue),
+    // et le mapping image→écran ne doit PAS bouger (tout le reste du compositeur en
+    // dépend). Une version antérieure de ce correctif protégeait si bien le second qu'elle
+    // annulait le premier dès que le focus n'était pas centré — d'où le balayage sur des
+    // focus décentrés dans les deux tests.
+
+    /// Boîte paddée (padding 50 % → `scale_frame` 0.8) dans une sortie carrée : le cas
+    /// plein cadre de l'issue.
+    const PADDED: [f32; 4] = [0.1, 0.1, 0.8, 0.8];
+
+    /// Les zooms d'un preset (`ZOOM_DEPTH_SCALES`, TS) et des focus réalistes — dont des
+    /// focus très décentrés, que le suivi de curseur produit en permanence.
+    const ZOOMS: [f32; 6] = [1.0, 1.25, 1.5, 1.8, 2.2, 3.5];
+    const FOCUSES: [[f32; 2]; 6] = [
+        [0.5, 0.5],
+        [0.3, 0.5],
+        [0.5, 0.8],
+        [0.15, 0.9],
+        [0.85, 0.2],
+        [0.0, 1.0],
+    ];
+
+    /// Le couple (boîte, coupe) réellement envoyé au GPU. `u_max`/`v_max` à 1 et pas de
+    /// crop : la coupe est donc directement en fractions d'image.
+    fn drawn(base: [f32; 4], zoom: f32, focus: [f32; 2]) -> ([f32; 4], [f32; 4]) {
+        let cut_ref = screen_source_rect(1.0, 1.0, None, zoom, focus);
+        let cut = screen_source_rect(1.0, 1.0, None, 1.0, focus);
+        (remap_box(base, cut_ref, cut), cut)
+    }
+
+    /// Où un point de l'image atterrit à l'écran, en fraction du CADRE.
+    fn on_screen(base: [f32; 4], zoom: f32, focus: [f32; 2], point: [f32; 2]) -> [f32; 2] {
+        let (dst, src) = drawn(base, zoom, focus);
+        let at = |f: f32, s0: f32, s1: f32, d0: f32, dw: f32| d0 + dw * (f - s0) / (s1 - s0);
+        [
+            at(point[0], src[0], src[2], dst[0], dst[2]),
+            at(point[1], src[1], src[3], dst[1], dst[3]),
+        ]
+    }
+
+    /// Le mapping d'avant : la coupe zoomée remplissait la boîte paddée, sans la bouger.
+    fn on_screen_before(base: [f32; 4], zoom: f32, focus: [f32; 2], point: [f32; 2]) -> [f32; 2] {
+        let src = screen_source_rect(1.0, 1.0, None, zoom, focus);
+        let at = |f: f32, s0: f32, s1: f32, d0: f32, dw: f32| d0 + dw * (f - s0) / (s1 - s0);
+        [
+            at(point[0], src[0], src[2], base[0], base[2]),
+            at(point[1], src[1], src[3], base[1], base[3]),
+        ]
+    }
+
+    /// L'invariant : rendre le zoom à la boîte ne déplace AUCUN point de l'image — même
+    /// grossissement, même cadrage. Seule l'étendue dessinée change.
+    #[test]
+    fn handing_the_zoom_to_the_box_moves_no_pixel() {
+        for &zoom in &ZOOMS {
+            for &focus in &FOCUSES {
+                for &point in &[[0.5, 0.5], [0.0, 0.0], [1.0, 1.0], [0.25, 0.75]] {
+                    let (was, now) = (
+                        on_screen_before(PADDED, zoom, focus, point),
+                        on_screen(PADDED, zoom, focus, point),
+                    );
+                    assert!(
+                        (was[0] - now[0]).abs() < 1e-4 && (was[1] - now[1]).abs() < 1e-4,
+                        "point {point:?} déplacé (zoom {zoom}, focus {focus:?}) : {was:?} → {now:?}"
+                    );
+                }
+            }
+        }
+    }
+
+    /// Ce que l'issue demande, et la régression que le testeur a vue : dès qu'on zoome, la
+    /// boîte doit déborder le rect paddé — y compris (surtout) avec un focus décentré.
+    #[test]
+    fn any_zoom_overflows_the_padding() {
+        for &zoom in &ZOOMS {
+            for &focus in &FOCUSES {
+                let (dst, _) = drawn(PADDED, zoom, focus);
+                let grew = dst[2] / PADDED[2];
+                assert!(
+                    (grew - zoom).abs() < 1e-4,
+                    "la boîte n'a pas pris le zoom (zoom {zoom}, focus {focus:?}) : ×{grew}"
+                );
+                if zoom > 1.0 {
+                    // Elle dépasse le rect paddé d'au moins un bord, donc mange du padding.
+                    assert!(
+                        dst[0] < PADDED[0] - 1e-6 || dst[0] + dst[2] > PADDED[0] + PADDED[2] + 1e-6,
+                        "boîte encore dans le padding (zoom {zoom}, focus {focus:?}) : {dst:?}"
+                    );
+                }
+            }
+        }
+        // Focus centré : le padding disparaît des QUATRE côtés dès que le zoom suffit à
+        // couvrir le cadre (ici 1/0.8 = 1.25).
+        let (dst, _) = drawn(PADDED, 1.25, [0.5, 0.5]);
+        assert_rect(dst, [0.0, 0.0, 1.0, 1.0]);
+        // Sans padding il n'y a rien à déborder, mais la boîte porte quand même le zoom.
+        let (dst, _) = drawn([0.0, 0.0, 1.0, 1.0], 2.0, [0.5, 0.5]);
+        assert_rect(dst, [-0.5, -0.5, 2.0, 2.0]);
+    }
+
+    // --- cover_crop_uv : la caméra n'est jamais étirée --------------------
+    // Le ratio de la coupe source, ramené en pixels d'image, doit TOUJOURS égaler
+    // celui de la boîte : c'est la définition de « pas de déformation ».
+
+    /// Ratio largeur/hauteur de la coupe, exprimé en pixels de l'image source.
+    fn crop_aspect(uv: (f32, f32, f32, f32), tex: [f32; 2]) -> f32 {
+        ((uv.2 - uv.0) * tex[0]) / ((uv.3 - uv.1) * tex[1])
+    }
+
+    /// L'invariant, balayé sur des boîtes très diverses — dont le slot en colonne
+    /// du preset side-by-side, qui est précisément le cas qui étirait la caméra.
+    #[test]
+    fn cover_crop_never_distorts_whatever_the_destination_box() {
+        let tex = [1024.0, 1024.0];
+        for &cam in &[[1280.0, 720.0], [960.0, 720.0], [640.0, 480.0]] {
+            for &box_ar in &[0.35, 0.5, 0.75, 1.0, 16.0 / 9.0, 2.4] {
+                let uv = cover_crop_uv(cam, tex, box_ar);
+                let got = crop_aspect(uv, tex);
+                assert!(
+                    (got - box_ar).abs() < 1e-3,
+                    "cam {cam:?} boite {box_ar} → coupe de ratio {got}, attendu {box_ar}",
+                );
+            }
+        }
+    }
+
+    /// La coupe reste DANS l'image visible et centrée — on ne va jamais chercher
+    /// le padding décodeur au-delà de `visible`, qui contient des pixels indéfinis.
+    #[test]
+    fn cover_crop_stays_inside_the_visible_frame_and_is_centred() {
+        let (cam, tex) = ([1280.0, 720.0], [2048.0, 1024.0]);
+        for &box_ar in &[0.35, 1.0, 2.4] {
+            let (u0, v0, u1, v1) = cover_crop_uv(cam, tex, box_ar);
+            assert!(u0 >= 0.0 && v0 >= 0.0, "coupe hors image: {u0},{v0}");
+            assert!(u1 <= cam[0] / tex[0] + 1e-6, "u1 {u1} deborde la largeur visible");
+            assert!(v1 <= cam[1] / tex[1] + 1e-6, "v1 {v1} deborde la hauteur visible");
+            let (mx, my) = (u0 + u1, v0 + v1);
+            assert!((mx - cam[0] / tex[0]).abs() < 1e-6, "pas centre en x");
+            assert!((my - cam[1] / tex[1]).abs() < 1e-6, "pas centre en y");
+        }
+    }
+
+    /// L'écran en layout bloc : le cover s'applique au rect DÉJÀ réduit par le crop
+    /// et le zoom. Quel que soit ce rect de départ, ce qui atterrit dans la boîte a
+    /// le ratio de la boîte — c'est ce qui empêche l'étirement.
+    #[test]
+    fn cover_uv_rect_gives_the_box_aspect_whatever_the_crop_and_zoom_left() {
+        let tex = [2048.0, 1024.0];
+        // rects source plausibles : plein cadre, bande verticale (crop portrait), zoom serré
+        for &uv in &[
+            [0.0, 0.0, 0.9375, 0.7031],
+            [0.41, 0.04, 0.55, 0.67],
+            [0.30, 0.20, 0.55, 0.45],
+        ] {
+            for &box_ar in &[0.4, 0.75, 1.0, 1.9, 3.2] {
+                let out = cover_uv_rect(uv, tex, box_ar);
+                let got = ((out[2] - out[0]) * tex[0]) / ((out[3] - out[1]) * tex[1]);
+                assert!(
+                    (got - box_ar).abs() / box_ar < 1e-3,
+                    "uv {uv:?} boite {box_ar} -> ratio {got}",
+                );
+                // le cover RÉDUIT : il ne va jamais chercher des pixels hors du rect source
+                assert!(out[0] >= uv[0] - 1e-6 && out[1] >= uv[1] - 1e-6, "deborde en haut/gauche");
+                assert!(out[2] <= uv[2] + 1e-6 && out[3] <= uv[3] + 1e-6, "deborde en bas/droite");
+            }
+        }
+    }
+
+    /// Propriété de sûreté : quand la boîte a DÉJÀ le ratio de la source (tous les
+    /// placements qui étaient corrects — PiP par défaut, vertical-stack, et le
+    /// center-crop carré de square/circle), la coupe est la frame entière. Le
+    /// correctif ne peut donc pas déplacer un pixel de ces cas-là.
+    #[test]
+    fn cover_crop_is_the_whole_frame_when_the_box_already_matches() {
+        let (cam, tex) = ([1280.0, 720.0], [2048.0, 1024.0]);
+        let uv = cover_crop_uv(cam, tex, cam[0] / cam[1]);
+        assert!((uv.0).abs() < 1e-6 && (uv.1).abs() < 1e-6);
+        assert!((uv.2 - cam[0] / tex[0]).abs() < 1e-6);
+        assert!((uv.3 - cam[1] / tex[1]).abs() < 1e-6);
+        // et une boîte carrée sur une source 4:3 redonne bien le center-crop carré
+        // que l'ancien branchement `is_square_shape` codait à la main.
+        let (su0, _, su1, _) = cover_crop_uv([960.0, 720.0], tex, 1.0);
+        assert!((su0 - (960.0 - 720.0) * 0.5 / tex[0]).abs() < 1e-6);
+        assert!((su1 - (960.0 + 720.0) * 0.5 / tex[0]).abs() < 1e-6);
+    }
+}
